:truck: rename: api 요청 코드를 용도에 따라 분류하고 기존 파일 삭제
